backend/requirements.txt
fastapi==0.94.1
fastapi-cache2==0.2.1
loguru==0.6.0
pydantic==1.10.6
python-decouple==3.8
python-jose==3.3.0
redis==4.5.3
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.40.0
uvicorn==0.21.1
PyMySQL==1.0.2
gunicorn
cryptography
python-multipart